rest controller enhancements and full integration with mongodb repository

+++ src/main/java/com/mtt/guildhome/guildpost/springrest/GuildPostRestController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,47 +13,88 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.mtt.guildhome.guildpost.domain.GuildPost;
+import com.mtt.guildhome.guildpost.domain.GuildPostFactory;
 import com.mtt.guildhome.guildpost.domain.GuildPostRepository;
 
 @RestController
-@RequestMapping("/guild/post")
+@RequestMapping("/v1/guilds/{guildId}/posts")
 public class GuildPostRestController {
 
 	@Autowired
 	private GuildPostRepository guildPostRepository;
 
 	@GetMapping("/{id}")
-	public GuildPost getPost(@PathVariable String id) {
-		return this.guildPostRepository.findById(id);
+	public ResponseEntity<?> getPost(@PathVariable String guildId, @PathVariable String id) {
+		GuildPost post = this.guildPostRepository.findById(id);
+
+		if (post == null) {
+			return new ResponseEntity<>("id-" + id, HttpStatus.NOT_FOUND);
+		}
+		if (!post.getGuildId().equalsIgnoreCase(guildId)) {
+			return ResponseEntity.badRequest().body("guildId mismatch");
+		}
+		return ResponseEntity.ok(post);
 	}
 
 	@GetMapping
-	public List<GuildPost> getPosts(@RequestParam String guildId, @RequestParam(required = false) String userId) {
+	public List<GuildPost> getPostsForGuild(@PathVariable String guildId,
+			@RequestParam(required = false) String userId) {
+		// TODO: pagination
 		if (userId == null) {
 			return this.guildPostRepository.findByGuild(guildId);
 		}
-
 		return this.guildPostRepository.findByGuildAndUser(guildId, userId);
 	}
 
 	@PostMapping
-	public void createPost(@RequestBody GuildPost post) {
+	public ResponseEntity<?> createPost(@PathVariable String guildId, @RequestBody GuildPost post) {
+		if (!guildId.equalsIgnoreCase(post.getGuildId())) {
+			return ResponseEntity.badRequest().body("guildId mismatch");
+		}
+
+		post = GuildPostFactory.newPost(guildId, post.getUserId(), post.getContent());
 		this.guildPostRepository.save(post);
+		return ResponseEntity.ok(post);
 	}
 
-	@PutMapping
-	public void updatePost(@RequestBody GuildPost post) {
+	@PutMapping("/{id}")
+	public ResponseEntity<?> updatePost(@PathVariable String guildId, @PathVariable String id,
+			@RequestBody GuildPost post) {
+		if (!guildId.equalsIgnoreCase(post.getGuildId()) || !id.equalsIgnoreCase(post.getId())) {
+			return ResponseEntity.badRequest().body("guildId/id mismatch");
+		}
+
+		GuildPost existingPost = this.guildPostRepository.findById(id);
+
+		if (existingPost == null) {
+			return new ResponseEntity<>("id-" + id, HttpStatus.NOT_FOUND);
+		}
+		if (!existingPost.getGuildId().equalsIgnoreCase(guildId)) {
+			return ResponseEntity.badRequest().body("guildId mismatch");
+		}
+
+		post = GuildPostFactory.updatedPost(existingPost.getId(), existingPost.getGuildId(), existingPost.getUserId(),
+				post.getContent());
 		this.guildPostRepository.save(post);
+		return ResponseEntity.ok(post);
 	}
 
 	@DeleteMapping("/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void deletePost(@PathVariable String id) {
+	public ResponseEntity<?> deletePost(@PathVariable String guildId, @PathVariable String id) {
+		GuildPost post = this.guildPostRepository.findById(id);
+
+		if (post == null) {
+			return new ResponseEntity<>("id-" + id, HttpStatus.NOT_FOUND);
+		}
+		if (!post.getGuildId().equalsIgnoreCase(guildId)) {
+			return ResponseEntity.badRequest().body("guildId mismatch");
+		}
+
 		this.guildPostRepository.delete(id);
+		return ResponseEntity.noContent().build();
 	}
 
 }
